added takeout-system service tests

+++ src/main/java/com/takeout/takeout_system/services/ProcessOrderService.java
@@ -2,11 +2,12 @@ package com.takeout.takeout_system.services;
 
 import com.takeout.takeout_system.data.dto.EnterItemRequest;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public interface ProcessOrderService {
     Boolean makeNewOrder();
     Boolean enterItem(EnterItemRequest enterItemRequest);
     BigInteger endOrder();
-    Boolean makeCashPayment(BigInteger amount);
+    Boolean makeCashPayment(BigDecimal amount);
 }
